Finished protection on categories.php

+++ admin/categories.php
@@ -9,33 +9,35 @@ $rows = get_category_array();
 
 if (count($rows) > 0){
 	foreach($rows as $row){
-		$temp = new category($row["id"]);
-		
-		$table_rows .= '
-			<tr>
-				<td align="center" style="width: 15ex; white-space: nowrap;">
-					<nobr>';
-		
-		if ($admin_user->can_admin($row["id"])){
+		if ($view_user->can_view($row["id"]) || $admin_user->can_admin($row["id"])){
+			$temp = new category($row["id"]);
+			
 			$table_rows .= '
-						[<a href="edit_category.php?id='.$row["id"].'">edit</a>]
-						[<a href="delete_category.php?id='.$row["id"].'">delete</a>]';
+				<tr>
+					<td align="center" style="width: 15ex; white-space: nowrap;">
+						<nobr>';
+			
+			if ($admin_user->can_admin($row["id"])){
+				$table_rows .= '
+							[<a href="edit_category.php?id='.$row["id"].'">edit</a>]
+							[<a href="delete_category.php?id='.$row["id"].'">delete</a>]';
+			}
+			
+			$table_rows .= '
+						</nobr>
+					</td>
+					<td style="white-space: nowrap;">';
+			
+			if ($view_user->can_view($row["id"])){
+				$table_rows .= $temp->get_breadcrumb_links();
+			}
+			else{
+				$table_rows .= $temp->breadcrumb_names;
+			}
+			
+			$table_rows .= ' ('.$temp->num_items_r().')</td>
+				</tr>';
 		}
-		
-		$table_rows .= '
-					</nobr>
-				</td>
-				<td style="white-space: nowrap;">';
-		
-		if ($view_user->can_view($row["id"])){
-			$table_rows .= $temp->get_breadcrumb_links();
-		}
-		else{
-			$table_rows .= $temp->breadcrumb_names;
-		}
-		
-		$table_rows .= ' ('.$temp->num_items_r().')</td>
-			</tr>';
 	}
 	
 	$table_rows = '<table>'.$table_rows.'</table>';
